HUGE UI updates to social history
